fix: move shared modules into the functions folder
