+++ src/components/machines/index.js
@@ -8,7 +8,7 @@ import {
   createProjectFormData,
   formatDegrees,
   roundAccurately,
-} from '../helpers/index.mjs';
+} from '../../../functions/shared/index.js';
 
 export const updateContext = (context, field, value) => {
   if (!field) {
