+++ src/components/pageElements/AddPoint.jsx
@@ -12,7 +12,7 @@ import { Switch } from '../formElements/Switch.jsx';
 import Spacer from '../formElements/Spacer.jsx';
 import { NumberedForm, NumberedFormSection } from '../formElements/Form.jsx';
 import Card from '../formElements/Card.jsx';
-import { addPointSchema as schema } from './CornerSubmission/Schema.mjs';
+import { addPointSchema as schema } from '../../../functions/shared/cornerSubmission/Schema.js';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { ErrorMessage } from '@hookform/error-message';
 import ErrorMessageTag from './ErrorMessage.jsx';
