+++ src/components/pageElements/Profile.jsx
@@ -7,7 +7,7 @@ import { Controller, useForm, useWatch } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { ErrorMessage } from '@hookform/error-message';
 import ErrorMessageTag from './ErrorMessage.jsx';
-import { profileSchema as schema } from './CornerSubmission/Schema.mjs';
+import { profileSchema as schema } from '../../../functions/shared/cornerSubmission/Schema.js';
 import { Input, Label } from '../formElements/Inputs.jsx';
 import Card from '../formElements/Card.jsx';
 import { Button } from '../formElements/Buttons.jsx';
